Break source code into modules
