+++ .gitignore
@@ -122,8 +122,3 @@ out
 
 tags
 tags.*
-
-schedule.etag
-schedule.json
-security.etag
-security.json
