+++ setup-test.js
@@ -1,8 +1,8 @@
-const isOldEnough = require('./eol-versions')
+const isNodeEOL = require('./src/is-node-eol')
 const assert = require('assert')
 
 // When old enough an error is thrown
-if (isOldEnough(process.version)) {
+if (isNodeEOL(process.version)) {
   runCompatibilityTest()
 } else {
   require('./test')
@@ -11,7 +11,7 @@ if (isOldEnough(process.version)) {
 function runCompatibilityTest () {
   const childProcess = require('child_process')
   const path = require('path')
-  const isNodeVulnerablePath = path.resolve('./index.js')
+  const isNodeVulnerablePath = path.resolve('./src/index.js')
   const child = childProcess.spawnSync(process.execPath, [isNodeVulnerablePath])
   assert.strictEqual(child.status, 1)
   assert(child.stdout.toString().indexOf('is end-of-life. There are high chances of being vulnerable') !== -1)
