+++ src/ascii.js
