+++ src/index.js
@@ -1,12 +1,12 @@
 #!/usr/bin/env node
 
-const isOldEnough = require('./eol-versions')
+const isNodeEOL = require('./is-node-eol')
 const ascii = require('./ascii')
 
 // To guarantee support on older versions and do not drastically impact
 // the maintenance of this module, we check if process.version is too old
 // and throw EOL warning when true.
-if (isOldEnough(process.version)) {
+if (isNodeEOL(process.version)) {
   console.log(ascii.danger)
   const msg = process.version + ' is end-of-life. There are high chances of being vulnerable. Please upgrade it.'
   console.log(msg)
@@ -14,10 +14,10 @@ if (isOldEnough(process.version)) {
 } else {
   // CLI
   if (require.main === module) {
-    require('./is-vulnerable').cli(process.version, require('os').platform())
+    require('./cli')(process.version, require('os').platform())
   } else {
     module.exports = {
-      isNodeVulnerable: require('./is-vulnerable').isNodeVulnerable
+      isNodeVulnerable: require('./is-node-vulnerable')
     }
   }
 }
