+++ src/utils/check-platform.js
@@ -1,0 +1,8 @@
+function checkPlatform (platform) {
+  const availablePlatforms = ['aix', 'darwin', 'freebsd', 'linux', 'openbsd', 'sunos', 'win32', 'android']
+  if (platform && !availablePlatforms.includes(platform)) {
+    throw new Error(`platform ${platform} is not valid. Please use ${availablePlatforms.join(',')}.`)
+  }
+}
+
+module.exports = checkPlatform
