+++ src/utils/debug.js
@@ -1,0 +1,7 @@
+function debug (msg) {
+  if (process.env.DEBUG) {
+    console.debug(msg)
+  }
+}
+
+module.exports = debug
