+++ src/utils/get-vulnerability-list.js
@@ -1,0 +1,23 @@
+const { bold } = require('../ascii')
+const isSystemAffected = require('./is-system-affected')
+const satisfies = require('semver/functions/satisfies')
+
+function getVulnerabilityList (currentVersion, data, platform) {
+  const list = []
+  for (const key in data) {
+    const vuln = data[key]
+
+    if (
+      (
+        satisfies(currentVersion, vuln.vulnerable) &&
+          !satisfies(currentVersion, vuln.patched)
+      ) && isSystemAffected(platform, vuln.affectedEnvironments)
+    ) {
+      const severity = vuln.severity === 'unknown' ? '' : `(${vuln.severity})`
+      list.push(`${bold(vuln.cve)}${severity}: ${vuln.overview}\n${bold('Patched versions')}: ${vuln.patched}`)
+    }
+  }
+  return list
+}
+
+module.exports = getVulnerabilityList
