+++ src/utils/is-system-affected.js
@@ -1,0 +1,14 @@
+const isSystemAffected = (platform, affectedEnvironments) => {
+  // No platform specified (legacy mode)
+  if (!platform || !Array.isArray(affectedEnvironments)) {
+    return true
+  }
+  // If the environment is matching or all the environments are affected
+  if (affectedEnvironments.includes(platform) || affectedEnvironments.includes('all')) {
+    return true
+  }
+  // Default to false
+  return false
+}
+
+module.exports = isSystemAffected
